requirements.txt
streamlit
pytesseract
opencv-python
Pillow
googletrans==4.0.0-rc1
numpy
scikit-learn